phase 1: addressing h bottleneck

+++ packboost/core.py
@@ -175,8 +175,8 @@ class PackBoost(BaseEstimator, RegressorMixin):
 
         # --- GPU fast path ---
         if (XS.is_cuda or Y.is_cuda or LF.is_cuda) and torch.cuda.is_available():
-            # Expect your compiled extension to expose kernels.h_sm
-            return kernels.h_sm(XS.contiguous(), Y.to(torch.int32).contiguous(), LF.contiguous(), int(max_depth))
+            # Expect your compiled extension to expose kernels.h_sm .to(torch.int32)
+            return kernels.h_sm(XS.contiguous(), Y.contiguous(), LF.contiguous(), int(max_depth))
 
         # --- CPU vectorized fallback ---
         dev = XS.device
